+++ app/manual-ledger/get-manual-ledger-requests.js
@@ -1,0 +1,24 @@
+const db = require('../data')
+
+const getManualLedgerRequests = async (paymentRequestId) => {
+  return db.manualLedgerPaymentRequest.findAll({
+    include: [{
+      model: db.paymentRequest,
+      as: 'ledgerPaymentRequest',
+      include: [{
+        model: db.invoiceLine,
+        as: 'invoiceLines'
+      },
+      {
+        model: db.scheme,
+        as: 'schemes'
+      }]
+    }],
+    where: {
+      paymentRequestId,
+      active: true
+    }
+  })
+}
+
+module.exports = getManualLedgerRequests
